// app/interfaces/create-tutor.ts
export interface CreateTutor {
    email: string;
    name: string;
    lastName: string;
    dni: string;
    birthdate: string;
    sex: string;
    telephone: string;
    yearEntry: string;
    observations?: string;
    category: string;
    countryId: number;
    dedication: string;
    dedicationDays: number;
    departmentId: number;
}